Add Validator to decouple TxController from Byron
Previously lots of Byron-specific logic was inside of TxController to do
some additional validation in order to have better error reporting upon
tx sending. This refactors this into a specific Validator interface that
does this checking in preparation for Shelley where we could make a
ShelleyValidator.
