+++ src/interfaces/index.js
@@ -8,3 +8,4 @@ export { Logger } from './logger'
 export { Genesis } from './genesis'
 export { NetworkConfig } from './network-config'
 export { StorageProcessor } from './storage-processor'
+export { Validator } from './validator'
