+++ src/interfaces/validator.js
@@ -1,0 +1,5 @@
+// @flow
+
+export interface Validator {
+    validateTx(txObj): any;
+}
